backend-vps/pyproject.toml
[build-system]
requires = ["setuptools>=66.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "study-agents"
version = "0.1.0"
description = "Local agent stack for PDF RAG + vision-driven subject-matter-expert assistance"
requires-python = ">=3.10"
dependencies = [
  "aiohttp",
  "beautifulsoup4",
  "numpy",
  "openai",
  "ollama",
  "python-dotenv",
  "requests",
  "supabase",
  "mcp[cli]",
]

[project.optional-dependencies]
api = [
  "fastapi",
  "uvicorn",
  "pydantic-ai",
]
ocr = [
  "docling[rapidocr]",
  "pymupdf",
  "Pillow",
  "pytesseract",
  "rapidocr-onnxruntime",
]
cag = [
  "fastapi",
  "uvicorn",
  "docling[rapidocr]",
  "pymupdf",
  "Pillow",
  "pytesseract",
  "rapidocr-onnxruntime",
]
rag = [
  "aiohttp",
  "docling[rapidocr]",
  "pymupdf",
  "Pillow",
  "pytesseract",
  "rapidocr-onnxruntime",
]
copilot = [
  "fastapi",
  "uvicorn",
  "pydantic-ai",
  "docling[rapidocr]",
  "pymupdf",
  "Pillow",
  "pytesseract",
  "rapidocr-onnxruntime",
  "mss",
]
vision-client = [
  "docling[rapidocr]",
  "keyboard",
  "mss",
  "pdf2image",
  "Pillow",
  "qrcode[pil]",
  "pytesseract",
  "rapidocr-onnxruntime",
]
easyocr = [
  "easyocr<2.0,>=1.7",
]
server = [
  "fastapi",
  "uvicorn",
  "pydantic-ai",
  "docling[rapidocr]",
  "pymupdf",
  "Pillow",
  "pytesseract",
  "rapidocr-onnxruntime",
  "mss",
]
full = [
  "fastapi",
  "uvicorn",
  "pydantic-ai",
  "docling[rapidocr]",
  "pymupdf",
  "Pillow",
  "qrcode[pil]",
  "pytesseract",
  "rapidocr-onnxruntime",
  "keyboard",
  "mss",
  "pdf2image",
]

[project.scripts]
study-agents-mcp = "study_agents.mcp_server_fixed:main"
study-agents-rag = "study_agents.rag_agent_cli:main"
study-agents-web-research = "study_agents.web_research_agent:main"
study-agents-cag = "study_agents.cag_agent:main"
study-agents-vision = "study_agents.vision_agent:main_loop"
study-agents-graph-inspector = "study_agents.graph_inspector:main"
study-agents-api = "study_agents.api_server:main"
study-agents-kg-visualizer = "study_agents.kg_visualizer_agent:main"
study-agents-manage = "study_agents.orchestrator:main"
study-agents-domain-profile = "study_agents.domain_profile_agent:main"
study-agents-validate = "study_agents.settings:validate_cli"
study-agents-scenario-api = "study_agents.scenario_api:main"
